Added the possibility to observe property changes
1) StylyableProperty now is ObservableProperty.
2)Obersers can Subscribe to a ObservableProperty.

+++ Sgl/Base/Time/TimeSpan.h
@@ -66,7 +66,7 @@ namespace Sgl
 			return TimeSpan(0);
 		}
 
-		constexpr double ToNanoseconds() const noexcept
+		constexpr long long ToNanoseconds() const noexcept
 		{
 			return _nanoseconds;
 		}
